requirements.txt
gdax
python-binance
python-bittrex